Code cleanup
git-svn-id: https://svn.code.sf.net/p/benerator/code/benerator/trunk@3082 37d91da1-6759-5a44-b2bd-541c34f01409

+++ src/main/java/org/databene/benerator/sample/StateGenerator.java
@@ -75,7 +75,7 @@ public class StateGenerator<E> implements Generator<E> {
     	try {
     		WeightedTransition[] ts = BeneratorScriptParser.parseTransitionList(transitionSpec);
 	    	for (WeightedTransition t : ts)
-	    		addTransition((E) t.getFrom(), (E) t.getTo(), (Double) t.getWeight().evaluate(null));
+	    		addTransition((E) t.getFrom(), (E) t.getTo(), t.getWeight().evaluate(null));
     	} catch (ParseException e) {
     		throw new ConfigurationError("Error parsing state machine specification: " + transitionSpec, e);
         }
